Added Hunter's Aspect of the Cheetah
Includes a setting in the class options to activate it.
It is available from level 5, and is implemented so that it will only
trigger if you're moving (otherwise falling back to any mount).

Additionally I restructured two of the three main loops,
BSL:GroundMountButton and BSL:Regular, to check for this before
the NeedsChauffeur test (because it's available before riding training).